Rename and references updates
