+++ lib/airtap.js
@@ -6,9 +6,9 @@ var EventEmitter = require('events').EventEmitter
 
 var createControlServer = require('./control-app')
 var setup_test_instance = require('./setup')
-var SauceBrowser = require('./sauce-browserlist')
-var PhantomBrowser = require('./phantom-browserlist')
-var Electron = require('./electron-browserlist')
+var SauceBrowser = require('./sauce-browser')
+var PhantomBrowser = require('./phantom-browser')
+var ElectronBrowser = require('./electron-browser')
 
 module.exports = Zuul
 
@@ -98,7 +98,7 @@ Zuul.prototype.run = function (cb) {
     }
 
     if (config.electron) {
-      var electron = Electron(config)
+      var electron = ElectronBrowser(config)
       self.emit('browser', electron)
       electron.once('done', function (results) {
         exit(null, results.failed === 0 && results.passed > 0)
