+++ lib/electron-browser.js
