+++ lib/get-saucelabs-browsers.js
