+++ lib/phantom-browser.js
