+++ lib/sauce-browser.js
