+++ test/integration/tape-sauce.js
@@ -2,7 +2,7 @@ var test = require('tape')
 var Zuul = require('../../')
 var after = require('after')
 var auth = require('../auth')
-var getBrowsers = require('../../lib/scout-browserlist')
+var getBrowsers = require('../../lib/get-saucelabs-browsers')
 var flattenBrowser = require('../../lib/flatten-browserlist')
 var browsersToTest = require('airtap-browsers').pullRequest
 var verify = require('./verify-common')
